Rename to make more clear; put in email test code to send email in preview to me on deploy
