--- emails/TemplateConfig.ts
@@ -1,4 +1,0 @@
-export interface TemplateConfig<T> {
-  templateName: string;
-  locals: T;
-}
