+++ emails/TemplatedEmail.ts
@@ -1,0 +1,4 @@
+export interface TemplatedEmail<T extends Record<string, any>> {
+  templateName: string;
+  locals: T;
+}
