+++ src/pages/api/user.ts
@@ -6,6 +6,8 @@ import { SessionUser } from "&server/models/SessionUser";
 import { AuthenticationError } from "&server/utils/AuthenticationError";
 import { ADMIN_ROLE } from "&server/utils/Authentication";
 import { MetricReporter } from "&server/utils/MetricReporter";
+import { sendEmail } from "&server/utils/Email";
+import { StatusEmail } from "../../../emails/status/StatusEmail";
 
 const METRIC_REPORTER = new MetricReporter();
 const SOURCE_NAME = "User Route";
@@ -67,6 +69,15 @@ const handler = generateMethodRoute(
       routeConfiguration: { requiredRoles: [ADMIN_ROLE] },
     },
   }
+);
+
+// test to see if an email actually sends
+void sendEmail(
+  "navbarry@gmail.com",
+  new StatusEmail({
+    name: "Bryce",
+    status: 0,
+  })
 );
 
 function validateUserHasAccessToUser(
